Re #26957 Turn ConvertToQ into free function from Python class
Converts ConvertToQ into a python function rather than a class, as we
don't need to persist any state
