+++ Framework/PythonInterface/plugins/algorithms/WorkflowAlgorithms/SANS/SANSBeamCentreFinderCore.py
@@ -12,7 +12,7 @@ from __future__ import (absolute_import, division, print_function)
 from mantid.api import (DataProcessorAlgorithm, MatrixWorkspaceProperty, AlgorithmFactory, PropertyMode, Progress,
                         IEventWorkspace)
 from mantid.kernel import (Direction, PropertyManagerProperty, StringListValidator)
-from sans.algorithm_detail.ConvertToQ import ConvertToQ
+from sans.algorithm_detail.convert_to_q import convert_workspace
 from sans.common.constants import EMPTY_NAME
 from sans.common.general_functions import create_child_algorithm, append_to_sans_file_tag
 from sans.state.state_base import create_deserialized_sans_state_from_property_manager
@@ -403,11 +403,11 @@ class SANSBeamCentreFinderCore(DataProcessorAlgorithm):
         @param wavelength_and_pixel_adjustment_workspace: the wavelength and pixel adjustment workspace.
         @return: a reduced workspace
         """
-        alg = ConvertToQ(state_convert_to_q=state.convert_to_q,
-                         wavelength_adj_workspace=wavelength_adjustment_workspace,
-                         pixel_adj_workspace=pixel_adjustment_workspace,
-                         wavelength_and_pixel_adj_workspace=wavelength_and_pixel_adjustment_workspace)
-        output_dict = alg.convert_workspace(workspace=workspace, output_summed_parts=True)
+        output_dict = convert_workspace(workspace=workspace, state_convert_to_q=state.convert_to_q,
+                                        wavelength_adj_workspace=wavelength_adjustment_workspace,
+                                        pixel_adj_workspace=pixel_adjustment_workspace,
+                                        wavelength_and_pixel_adj_workspace=wavelength_and_pixel_adjustment_workspace,
+                                        output_summed_parts=True)
 
         output_workspace = output_dict["output"]
         sum_of_counts = output_dict["counts_summed"]
